Fix initialization order bug.
I introduced this bug in 7f1ce55c7b0ca40192c9e5e4f4804ef4860bf854, which was intended to be a cleanup.

The constructor which takes a `LockManager` as a parameter requires `name` to be initialized before `vfs`, so `name` must be declared first in the class.

The tests didn't catch this because the tests don't actually use this constructor, because the tests cover only the default `LockManager`. But, this quickly broke code that uses the two-param constructor (in the SRS PR).