Make sql-test.wd-test actually work.
It looks like this test had not actually been made to work with `workerd test` and was instead being run by starting a server and sending HTTP requests.

To make it work with the tent harness, we need to make TEST_TMPDIR available and direct the DO storage to that. Not too bad.